fix formatting, move disable_interrupts and enable_interrupts to interrupts file
