+++ kernel/src/interrupts.rs
@@ -1,8 +1,102 @@
+use core::arch::asm;
+
 use libfenix::{noop, read_addr, set_bit, write_addr, INTC};
 
 const INTC_ILR: u32 = 0x100;
 const INTC_SIR_IRQ: u32 = 0x40;
 const INTC_CONTROL: u32 = 0x48;
+
+static mut INTERRUPT_HANDLERS: &mut [fn(); 128] = &mut [noop; 128];
+
+#[no_mangle]
+fn handle_interrupt() {
+    let interrupt = current();
+    execute(interrupt);
+    clear();
+}
+
+pub fn enable_interrupt(interrupt: Interrupt, mode: Mode, priority: u8) {
+    let interrupt_number = interrupt as u32;
+
+    let addr = INTC + INTC_ILR + (4 * interrupt_number);
+    let enable_fiq = match mode {
+        Mode::IRQ => 0,
+        Mode::FIQ => 1,
+    };
+    let bank = match InterruptBank::new(interrupt_number) {
+        Some(bank) => bank,
+        None => return,
+    };
+
+    write_addr(addr, enable_fiq | (priority << 2) as u32);
+    set_bit(INTC + bank.get_mir() + 4, interrupt_number % 32);
+}
+
+pub fn register_handler(handler: fn(), interrupt: Interrupt) {
+    unsafe {
+        INTERRUPT_HANDLERS[interrupt as usize] = handler;
+    }
+}
+
+pub fn current() -> Option<Interrupt> {
+    let num = read_addr(INTC + INTC_SIR_IRQ) & 0x7F;
+
+    Interrupt::new(num)
+}
+
+pub fn execute(interrupt: Option<Interrupt>) {
+    if let Some(interrupt) = interrupt {
+        unsafe { INTERRUPT_HANDLERS[interrupt as usize]() }
+    }
+}
+
+pub fn clear() {
+    write_addr(INTC + INTC_CONTROL, 0x1);
+}
+
+pub fn enable_interrupts() -> u32 {
+    let cpsr: u32;
+    unsafe {
+        asm!("mrs {0}, cpsr", out(reg) cpsr);
+        asm!("msr cpsr_c, {0}", in(reg) cpsr & !0x80)
+    };
+
+    cpsr
+}
+
+pub fn disable_interrupts() -> u32 {
+    let cpsr: u32;
+    unsafe {
+        asm!("mrs {0}, cpsr", out(reg) cpsr);
+        asm!("msr cpsr_c, {0}", in(reg) cpsr | 0x80)
+    };
+
+    cpsr
+}
+
+pub fn restore_cpsr(cpsr: u32) {
+    unsafe { asm!("msr cpsr_c, {0}", in(reg) cpsr) };
+}
+
+pub fn enabled<F, T>(f: F) -> T
+where
+    F: FnOnce() -> T,
+{
+    let cpsr = enable_interrupts();
+    let result = f();
+    restore_cpsr(cpsr);
+    result
+}
+
+pub fn free<F, T>(f: F) -> T
+where
+    F: FnOnce() -> T,
+{
+    let cpsr = disable_interrupts();
+    let result = f();
+    restore_cpsr(cpsr);
+    result
+}
 
 pub enum Interrupt {
     I2C2INT = 30,
@@ -29,54 +123,6 @@ impl Interrupt {
             _ => None,
         }
     }
-}
-
-#[no_mangle]
-fn handle_interrupt() {
-    let interrupt = current();
-    execute(interrupt);
-    clear();
-}
-
-pub fn enable_interrupt(interrupt: Interrupt, mode: Mode, priority: u8) {
-    let interrupt_number = interrupt as u32;
-
-    let addr = INTC + INTC_ILR + (4 * interrupt_number);
-    let enable_fiq = match mode {
-        Mode::IRQ => 0,
-        Mode::FIQ => 1,
-    };
-    let bank = match InterruptBank::new(interrupt_number) {
-        Some(bank) => bank,
-        None => return,
-    };
-
-    write_addr(addr, enable_fiq | (priority << 2) as u32);
-    set_bit(INTC + bank.get_mir() + 4, interrupt_number % 32);
-}
-
-static mut INTERRUPT_HANDLERS: &mut [fn(); 128] = &mut [noop; 128];
-
-pub fn register_handler(handler: fn(), interrupt: Interrupt) {
-    unsafe {
-        INTERRUPT_HANDLERS[interrupt as usize] = handler;
-    }
-}
-
-pub fn current() -> Option<Interrupt> {
-    let num = read_addr(INTC + INTC_SIR_IRQ) & 0x7F;
-
-    Interrupt::new(num)
-}
-
-pub fn execute(interrupt: Option<Interrupt>) {
-    if let Some(interrupt) = interrupt {
-        unsafe { INTERRUPT_HANDLERS[interrupt as usize]() }
-    }
-}
-
-pub fn clear() {
-    write_addr(INTC + INTC_CONTROL, 0x1);
 }
 
 pub enum Mode {
